chore: rename the `ethermint` to `chibaclonk`
1. rename `ethermint` to `chiba-clonk`

2. fix the `make localnet-start`
