+++ app/app.go
@@ -128,10 +128,10 @@ func init() {
 		panic(err)
 	}
 
-	DefaultNodeHome = filepath.Join(userHomeDir, ".ethermintd")
+	DefaultNodeHome = filepath.Join(userHomeDir, ".chibaclonkd")
 }
 
-const appName = "ethermintd"
+const appName = "chibaclonkd"
 
 var (
 	// DefaultNodeHome default home directories for the application daemon
