+++ scripts/start-docker.sh
@@ -1,11 +1,11 @@
-#!/bin/bash
+#!/bin/sh
 
 echo "prepare genesis: Run validate-genesis to ensure everything worked and that the genesis file is setup correctly"
-./ethermintd validate-genesis --home /ethermint
+chibaclonkd validate-genesis --home /chibaclonk
 
-echo "starting ethermint node $ID in background ..."
-./ethermintd start \
---home /ethermint \
+echo "starting chibaclonk node $ID in background ..."
+chibaclonkd start \
+--home /chibaclonk \
 --keyring-backend test
 
 echo "started ethermint node"
